Move third-party migration detection policies to domain
Extract source detection and asmdef reference decision rules into Domain so Infrastructure keeps only JSON parsing and mutation adapters for the C4-11 follow-up boundary.
