+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationAsmdefReferenceRules.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: caed201926624f59baa57daff869d14d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
