+++ Packages/src/Editor/Domain/ThirdPartyToolMigrationSourceDetectionRules.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ef7a2e333cc4aea990bbac82fab7c90
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
